Split annotations
